@@ -1,5 +1,6 @@
 package com.gdmap.newscctv;
 
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.graphics.Color;
 import android.graphics.Point;
@@ -11,17 +12,26 @@ import android.text.TextUtils;
 import android.text.style.ForegroundColorSpan;
 import android.util.Log;
 import android.util.TypedValue;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.animation.BounceInterpolator;
 import android.view.animation.Interpolator;
+import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.amap.api.location.AMapLocation;
+import com.amap.api.location.AMapLocationClient;
+import com.amap.api.location.AMapLocationClientOption;
+import com.amap.api.location.AMapLocationListener;
 import com.amap.api.maps.AMap;
 import com.amap.api.maps.AMap.InfoWindowAdapter;
+import com.amap.api.maps.CameraUpdateFactory;
+import com.amap.api.maps.LocationSource;
 import com.amap.api.maps.MapView;
 import com.amap.api.maps.Projection;
 import com.amap.api.maps.model.BitmapDescriptorFactory;
@@ -29,20 +39,30 @@ import com.amap.api.maps.model.CameraPosition;
 import com.amap.api.maps.model.LatLng;
 import com.amap.api.maps.model.Marker;
 import com.amap.api.maps.model.MarkerOptions;
+import com.amap.api.maps.model.NavigateArrowOptions;
 import com.amap.api.maps.model.animation.Animation;
 import com.amap.api.maps.model.animation.TranslateAnimation;
+import com.amap.api.services.core.AMapException;
 import com.amap.api.services.core.LatLonPoint;
+import com.amap.api.services.core.PoiItem;
+import com.amap.api.services.core.SuggestionCity;
+import com.amap.api.services.geocoder.GeocodeQuery;
 import com.amap.api.services.geocoder.GeocodeResult;
 import com.amap.api.services.geocoder.GeocodeSearch;
 import com.amap.api.services.geocoder.RegeocodeAddress;
 import com.amap.api.services.geocoder.RegeocodeQuery;
 import com.amap.api.services.geocoder.RegeocodeResult;
+import com.amap.api.services.poisearch.PoiResult;
+import com.amap.api.services.poisearch.PoiSearch;
 
 import java.lang.reflect.Type;
+import java.util.List;
+
+import butterknife.BindView;
 
 public class MarkerClickActivity extends CheckPermissionsActivity implements View.OnClickListener,
         AMap.OnMarkerClickListener, AMap.OnMapClickListener, AMap.OnInfoWindowClickListener,
-        GeocodeSearch.OnGeocodeSearchListener, InfoWindowAdapter, AMap.OnCameraChangeListener {
+        GeocodeSearch.OnGeocodeSearchListener, InfoWindowAdapter, AMap.OnCameraChangeListener, LocationSource, AMapLocationListener {
 
     private MarkerOptions markerOption;
     private Marker screenMarker;
@@ -50,8 +70,16 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
     private MapView mapView;
     private GeocodeSearch mGeocodeSearch;
     private CameraPosition mCameraPosition;
-    private LatLng latlng = new LatLng(39.91746, 116.396481);
+//    private LatLng latlng = new LatLng(39.91746, 116.396481);
+    private LatLng curLatlng;
     private boolean startJump;
+    private OnLocationChangedListener mListener;
+    private AMapLocationClient mlocationClient;
+    private AMapLocationClientOption mLocationOption;
+    private ProgressDialog progDialog = null;
+
+    public EditText mSearchEdt;
+    private String keyWords;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -66,6 +94,22 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
         mapView = (MapView) findViewById(R.id.map);
         mapView.onCreate(savedInstanceState); // 此方法必须重写
         init();
+        mSearchEdt = findViewById(R.id.search_text);
+        mSearchEdt.setOnKeyListener(new View.OnKeyListener() {
+            @Override
+            public boolean onKey(View v, int keyCode, KeyEvent event) {
+                if (keyCode == KeyEvent.KEYCODE_ENTER && event.getAction() == KeyEvent.ACTION_DOWN) {
+                    String searchString = mSearchEdt.getText().toString();
+                    if (TextUtils.isEmpty(searchString)) {
+                        Toast.makeText(MarkerClickActivity.this,"请填入搜索内容",Toast.LENGTH_SHORT).show();
+                        return true;
+                    }
+                    search(searchString);
+                    return true;
+                }
+                return false;
+            }
+        });
     }
 
     /**
@@ -76,15 +120,31 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
         clearMap.setOnClickListener(this);
         Button resetMap = (Button) findViewById(R.id.resetMap);
         resetMap.setOnClickListener(this);
+        progDialog = new ProgressDialog(this);
         if (aMap == null) {
             aMap = mapView.getMap();
+            setUpMap();
         }
-        setUpMap();
+    }
+
+    public void search(String key){
+        InputMethodManager imm = (InputMethodManager)getSystemService(Context.INPUT_METHOD_SERVICE);
+        if(imm != null){
+            imm.hideSoftInputFromWindow(mSearchEdt.getWindowToken(), 0);
+        }
+        // name表示地址，第二个参数表示查询城市，中文或者中文全拼，citycode、adcode
+        GeocodeQuery query = new GeocodeQuery(key, "");
+        mGeocodeSearch.getFromLocationNameAsyn(query);
     }
 
     private void setUpMap() {
         mGeocodeSearch = new GeocodeSearch(getApplicationContext());
         mGeocodeSearch.setOnGeocodeSearchListener(this);
+
+        aMap.setLocationSource(this);
+        aMap.getUiSettings().setMyLocationButtonEnabled(true);// 设置默认定位按钮是否显示
+        aMap.setMyLocationEnabled(true);// 设置为true表示显示定位层并可触发定位，false表示隐藏定位层并不可触发定位，默认是false
+
         aMap.setOnMarkerClickListener(this);
         aMap.setOnMapClickListener(null);
         aMap.setOnInfoWindowClickListener(this);
@@ -98,6 +158,7 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
         aMap.setOnCameraChangeListener(this);
 
         aMap.setInfoWindowAdapter(this);
+
     }
 
     @Override
@@ -107,8 +168,8 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
 
     @Override
     public void onCameraChangeFinish(CameraPosition cameraPosition) {
-        aMap.setOnCameraChangeListener(null);
         mCameraPosition = cameraPosition;
+        showDialog();
         startJumpAnimation();
     }
 
@@ -176,25 +237,32 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
     @Override
     public void onRegeocodeSearched(RegeocodeResult regeocodeResult, int i) {
         //地图移动的监听
-        aMap.setOnCameraChangeListener(this);
-        RegeocodeAddress regeocodeAddress = regeocodeResult.getRegeocodeAddress();
-        screenMarker.setTitle(regeocodeResult.getRegeocodeAddress().getDistrict());
-        if(regeocodeAddress != null){
-            String city =
-                    //regeocodeAddress.getCity()
-                    // + regeocodeAddress.getDistrict()
-                    regeocodeAddress.getTownship()
-                            + regeocodeAddress.getStreetNumber().getStreet()
-                            + regeocodeAddress.getStreetNumber().getNumber()
-                            + regeocodeAddress.getBuilding();
-            screenMarker.setSnippet(city);
+        dismissDialog();
+        if(i == AMapException.CODE_AMAP_SUCCESS){
+            RegeocodeAddress regeocodeAddress = regeocodeResult.getRegeocodeAddress();
+            screenMarker.setTitle(regeocodeResult.getRegeocodeAddress().getDistrict());
+            if(regeocodeAddress != null){
+                String city =
+                        //regeocodeAddress.getCity()
+                        // + regeocodeAddress.getDistrict()
+                        regeocodeAddress.getTownship()
+                                + regeocodeAddress.getStreetNumber().getStreet()
+                                + regeocodeAddress.getStreetNumber().getNumber()
+                                + regeocodeAddress.getBuilding();
+                screenMarker.setSnippet(city);
+            }
+            screenMarker.showInfoWindow();
         }
-        screenMarker.showInfoWindow();
     }
 
     @Override
     public void onGeocodeSearched(GeocodeResult geocodeResult, int i) {
-        aMap.setOnCameraChangeListener(this);
+        if(i==1000){
+            if(geocodeResult.getGeocodeAddressList()!=null&&geocodeResult.getGeocodeAddressList().size()>0){
+                LatLng curLatlng = new LatLng(geocodeResult.getGeocodeAddressList().get(0).getLatLonPoint().getLatitude(),geocodeResult.getGeocodeAddressList().get(0).getLatLonPoint().getLongitude());
+                aMap.moveCamera(CameraUpdateFactory.newLatLngZoom(curLatlng, 16f));
+            }
+        }
     }
 
     /**
@@ -251,7 +319,8 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
             case R.id.resetMap:
                 if (aMap != null) {
                     aMap.clear();
-                    addMarkersToMap(latlng);
+//                    addMarkersToMap(latlng);
+                    addMarkersToMap(curLatlng);
                 }
                 break;
             default:
@@ -272,13 +341,13 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
                 fromResource(R.drawable.purple_pin)).draggable(true).title("标题").snippet("内容").position(latLng).anchor(0.5f, 0.5f));
         screenMarker.showInfoWindow();
         screenMarker.setPositionByPixels(point.x, point.y);
+        screenMarker.setZIndex(1);
     }
 
     /**
      * 屏幕中心marker 跳动
      */
     public void startJumpAnimation() {
-
         if (screenMarker != null) {
             //根据屏幕距离计算需要移动的目标点
             final LatLng latLng = screenMarker.getPosition();
@@ -302,9 +371,11 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
             animation.setAnimationListener(new Animation.AnimationListener() {
                 @Override
                 public void onAnimationStart() {
-                    LatLonPoint latLngPoint = new LatLonPoint(mCameraPosition.target.latitude, mCameraPosition.target.longitude);
-                    RegeocodeQuery regeocodeQuery = new RegeocodeQuery(latLngPoint, 1000, GeocodeSearch.AMAP);
-                    mGeocodeSearch.getFromLocationAsyn(regeocodeQuery);
+                    if(mCameraPosition!=null){
+                        LatLonPoint latLngPoint = new LatLonPoint(mCameraPosition.target.latitude, mCameraPosition.target.longitude);
+                        RegeocodeQuery regeocodeQuery = new RegeocodeQuery(latLngPoint, 200, GeocodeSearch.AMAP);
+                        mGeocodeSearch.getFromLocationAsyn(regeocodeQuery);
+                    }
                 }
 
                 @Override
@@ -364,5 +435,67 @@ public class MarkerClickActivity extends CheckPermissionsActivity implements Vie
     private static int dip2px(Context context, float dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * scale + 0.5f);
+    }
+
+    @Override
+    public void activate(OnLocationChangedListener onLocationChangedListener) {
+        mListener = onLocationChangedListener;
+        if (mlocationClient == null) {
+            mlocationClient = new AMapLocationClient(this);
+            mLocationOption = new AMapLocationClientOption();
+            //设置定位监听
+            mlocationClient.setLocationListener(this);
+            //设置为高精度定位模式
+            mLocationOption.setOnceLocation(true);
+            mLocationOption.setLocationMode(AMapLocationClientOption.AMapLocationMode.Hight_Accuracy);
+            //设置定位参数
+            mlocationClient.setLocationOption(mLocationOption);
+            // 此方法为每隔固定时间会发起一次定位请求，为了减少电量消耗或网络流量消耗，
+            // 注意设置合适的定位时间的间隔（最小间隔支持为2000ms），并且在合适时间调用stopLocation()方法来取消定位请求
+            // 在定位结束后，在合适的生命周期调用onDestroy()方法
+            // 在单次定位情况下，定位无论成功与否，都无需调用stopLocation()方法移除请求，定位sdk内部会移除
+            mlocationClient.startLocation();
+        }
+    }
+
+    @Override
+    public void deactivate() {
+        mListener = null;
+        if (mlocationClient != null) {
+            mlocationClient.stopLocation();
+            mlocationClient.onDestroy();
+        }
+        mlocationClient = null;
+    }
+
+    @Override
+    public void onLocationChanged(AMapLocation aMapLocation) {
+         if(mListener!=null&& aMapLocation != null){
+             if (aMapLocation != null
+                     && aMapLocation.getErrorCode() == 0) {
+                 mListener.onLocationChanged(aMapLocation);
+                 curLatlng = new LatLng(aMapLocation.getLatitude(), aMapLocation.getLongitude());
+                 aMap.moveCamera(CameraUpdateFactory.newLatLngZoom(curLatlng, 16f));
+             }
+         }
+    }
+
+    public void showDialog() {
+        progDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
+        progDialog.setIndeterminate(false);
+        progDialog.setCancelable(true);
+        progDialog.setMessage("正在加载...");
+        progDialog.show();
+    }
+
+    public void dismissDialog() {
+        if (progDialog != null) {
+            progDialog.dismiss();
+        }
+    }
+
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
     }
 }
